Dissolve api.ts — inline each fetcher into its hook
