+++ projects/packages/podcast/src/dashboard/episodes/use-episodes-query.ts
@@ -1,5 +1,62 @@
 import { useQuery, keepPreviousData } from '@tanstack/react-query';
-import { fetchEpisodes, type EpisodesQueryArgs, type EpisodesPage } from '../api';
+import apiFetch from '@wordpress/api-fetch';
+import { addQueryArgs } from '@wordpress/url';
+import type { Episode } from '../types';
+
+export interface EpisodesQueryArgs {
+	categoryId: number;
+	page?: number;
+	perPage?: number;
+	orderBy?: 'date' | 'title';
+	order?: 'asc' | 'desc';
+	search?: string;
+	status?: string;
+}
+
+export interface EpisodesPage {
+	episodes: Episode[];
+	totalPages: number;
+	total: number;
+}
+
+const fetchEpisodes = async ( args: EpisodesQueryArgs ): Promise< EpisodesPage > => {
+	const {
+		categoryId,
+		page = 1,
+		perPage = 20,
+		orderBy = 'date',
+		order = 'desc',
+		search = '',
+		status = 'any',
+	} = args;
+
+	const query: Record< string, string | number > = {
+		categories: categoryId,
+		page,
+		per_page: perPage,
+		orderby: orderBy,
+		order,
+		_embed: 'wp:featuredmedia',
+	};
+	if ( search ) {
+		query.search = search;
+	}
+	if ( status ) {
+		query.status = status;
+	}
+
+	const response = ( await apiFetch( {
+		path: addQueryArgs( '/wp/v2/posts', query ),
+		method: 'GET',
+		parse: false,
+	} ) ) as Response;
+
+	const episodes = ( await response.json() ) as Episode[];
+	const total = parseInt( response.headers.get( 'X-WP-Total' ) || '0', 10 );
+	const totalPages = parseInt( response.headers.get( 'X-WP-TotalPages' ) || '1', 10 );
+
+	return { episodes, total, totalPages };
+};
 
 /**
  * Read a page of podcast episodes; keeps the previous page visible during
